GAでエラーが出ないように修正
