+++ app/about/page.tsx
@@ -1,3 +1,4 @@
+import { GA } from "@/ui/GA";
 import { InfoDisplay } from "@/ui/InfoDisplay";
 import { ReadMoreButton } from "@/ui/ReadMoreButton";
 import { SectionTitle } from "@/ui/SectionTitle";
@@ -10,6 +11,7 @@ import Link from 'next/link'
 export default function About() {
   return (
     <main className="min-h-screen bg-cyan-50 dark:bg-slate-500">
+      <GA/>
       <div className='md:flex justify-start md:px-24 lg:px-40 xl:px-[300px] pt-48'>
         <SectionTitle title='About'>私について</SectionTitle>
       </div>
