+++ app/layout.tsx
@@ -1,7 +1,5 @@
 import './globals.css';
 
-export const GA_TRACKING_ID = process.env.NEXT_PUBLIC_GA_ID || '';
-
 import type { Metadata } from 'next';
 import { Noto_Sans_JP } from 'next/font/google';
 
@@ -22,19 +20,6 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="ja">
-      <script async src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`} />
-      <script
-                dangerouslySetInnerHTML={{
-                  __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', ${GA_TRACKING_ID}, {
-              page_path: window.location.pathname,
-            });
-        `,
-                }}
-              />
       <body className={`${notoSansJP.className}`}>
         <Header />
         {children}
