+++ app/works/poster/page.tsx
@@ -1,3 +1,4 @@
+import { GA } from '@/ui/GA'
 import { InfoDisplay } from '@/ui/InfoDisplay'
 import { ReadMoreButton } from '@/ui/ReadMoreButton'
 import { SectionTitle } from '@/ui/SectionTitle'
@@ -9,6 +10,7 @@ import Link from 'next/link'
 export default function Poster() {
   return (
     <main className="min-h-screen bg-cyan-50 dark:bg-slate-500">
+      <GA/>
       <div className='md:flex justify-start md:px-24 lg:px-40 xl:px-[300px] pt-64'>
         <SectionTitle title='Poster'>成果ポスター</SectionTitle>
       </div>
