+++ ui/GA.tsx
@@ -1,0 +1,22 @@
+"use client";
+export const GA_TRACKING_ID = process.env.NEXT_PUBLIC_GA_ID || '';
+
+export function GA() {
+  return (
+    <>
+    <script async src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`} />
+    <script
+      dangerouslySetInnerHTML={{
+        __html: `
+          window.dataLayer = window.dataLayer || [];
+          function gtag(){dataLayer.push(arguments);}
+          gtag('js', new Date());
+          gtag('config', ${GA_TRACKING_ID}, {
+            page_path: window.location.pathname,
+          });
+      `,
+      }}
+      />
+      </>
+  );
+}
